changed rle fileops to binary, fread/fwrite now
Yeah, so... I've been running file ops in text mode using fputc/fgetc.
Doesn't really work on binary files. It still corrupts 'em, but it's better - the .bmp is still a .bmp even decompressed, and of correct height and width.
So, progress, even if not all the way there, yet. I have a 1024 byte buffer I use to read/write, and there's fiddly bits in the joins, I think.